feat: progress update - backend running successfully

+++ services/api/src/main/java/com/chubini/pku/products/FoodProductUpsertDto.java
@@ -1,12 +1,13 @@
 package com.chubini.pku.api.products;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record FoodProductUpsertDto(
     @NotBlank String name,
     @NotBlank String unit,
-    @NotNull @DecimalMin("0.0") Double proteinPer100g,
-    @NotNull @DecimalMin("0.0") Double phePer100g,
+    @NotNull @DecimalMin("0.0") BigDecimal proteinPer100g, // ✅
+    @NotNull @DecimalMin("0.0") BigDecimal phePer100g,      // ✅
     @NotNull @Min(0) Integer kcalPer100g,
     String category,
     Boolean isActive
